use cedar.Chart() instead of Cedar()
@esri/cedar exports a single global namesapce, cedar
the old Cedar() constructor is now cedar.Chart() and is no longer global
